pstats tracking for texture and vertex data on main memory

+++ panda/src/pstatclient/pStatClient.h
@@ -124,9 +124,9 @@ private:
   void stop(int collector_index, int thread_index, float as_of);
 
   void clear_level(int collector_index, int thread_index);
-  void set_level(int collector_index, int thread_index, float level);
-  void add_level(int collector_index, int thread_index, float increment);
-  float get_level(int collector_index, int thread_index) const;
+  void set_level(int collector_index, int thread_index, double level);
+  void add_level(int collector_index, int thread_index, double increment);
+  double get_level(int collector_index, int thread_index) const;
 
   static void start_clock_wait();
   static void start_clock_busy_wait();
@@ -155,7 +155,7 @@ private:
   public:
     PerThreadData();
     bool _has_level;
-    float _level;
+    double _level;
     int _nested_count;
   };
   typedef pvector<PerThreadData> PerThread;
